+++ examples/cpp_api/quickstart_map.cc
@@ -58,11 +58,12 @@ void write_map() {
   tiledb::Context ctx;
 
   // Open the map
-  tiledb::Map map(ctx, map_name);
+  tiledb::Map map(ctx, map_name, TILEDB_WRITE);
 
   map["key_1"]["a"] = 1;
   map["key_2"] = 2;  // Implicit "a" since there is 1 attr
   map["key_3"] = 3;
+  map.flush();
 
   // Close the map
   map.close();
@@ -72,7 +73,7 @@ void read_map() {
   Context ctx;
 
   // Open the map
-  tiledb::Map map(ctx, map_name);
+  tiledb::Map map(ctx, map_name, TILEDB_READ);
 
   // Read the keys
   int a1 = map["key_1"];