+++ test/src/unit-capi-kv.cc
@@ -91,14 +91,11 @@ struct KVFx {
   KVFx();
   ~KVFx();
   void check_single_read(const std::string& path);
-  void check_read_on_attribute_subset(const std::string& path);
   void check_iter(const std::string& path);
   void check_kv_item();
   void check_kv_item(tiledb_kv_item_t* kv_item);
-  void check_interleaved_read_write(const std::string& path);
   void check_write(const std::string& path);
   void create_kv(const std::string& path);
-  void check_kv_iter_with_dirty_kv(const std::string& path);
   void create_temp_dir(const std::string& path);
   void remove_temp_dir(const std::string& path);
   static std::string random_bucket_name(const std::string& prefix);
@@ -314,7 +311,7 @@ void KVFx::check_write(const std::string& path) {
   tiledb_kv_t* kv;
   int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
   REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, nullptr, 0);
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_WRITE);
   REQUIRE(rc == TILEDB_OK);
 
   // The kv is not dirty
@@ -450,6 +447,10 @@ void KVFx::check_write(const std::string& path) {
   rc = tiledb_kv_add_item(ctx_, kv, kv_item6);
   CHECK(rc == TILEDB_ERR);
 
+  // Flush
+  rc = tiledb_kv_flush(ctx_, kv);
+  REQUIRE(rc == TILEDB_OK);
+
   // Close kv
   rc = tiledb_kv_close(ctx_, kv);
   REQUIRE(rc == TILEDB_OK);
@@ -472,7 +473,6 @@ void KVFx::check_write(const std::string& path) {
 
 void KVFx::check_single_read(const std::string& path) {
   // Open key-value store
-  const char* attributes[] = {ATTR_1.c_str(), ATTR_2, ATTR_3};
   tiledb_kv_t* kv;
   int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
   REQUIRE(rc == TILEDB_OK);
@@ -481,7 +481,7 @@ void KVFx::check_single_read(const std::string& path) {
   rc = tiledb_kv_reopen(ctx_, kv);
   REQUIRE(rc == TILEDB_ERR);
 
-  rc = tiledb_kv_open(ctx_, kv, attributes, 3);
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_READ);
   REQUIRE(rc == TILEDB_OK);
 
   // Re-open should succeed after open
@@ -634,205 +634,6 @@ void KVFx::check_single_read(const std::string& path) {
   tiledb_kv_item_free(&kv_item4);
 }
 
-void KVFx::check_read_on_attribute_subset(const std::string& path) {
-  // Open key-value store
-  const char* attributes[] = {ATTR_1.c_str()};
-  tiledb_kv_t* kv;
-  int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, attributes, 1);
-  REQUIRE(rc == TILEDB_OK);
-
-  // For retrieving values
-  const void *a1, *a2;
-  uint64_t a1_size, a2_size;
-  tiledb_datatype_t a1_type, a2_type;
-
-  // Prepare keys
-  int key1 = 100;
-  tiledb_datatype_t key1_type = TILEDB_INT32;
-  uint64_t key1_size = sizeof(int);
-
-  float key2 = 200.0;
-  tiledb_datatype_t key2_type = TILEDB_FLOAT32;
-  uint64_t key2_size = sizeof(float);
-
-  double key3[] = {300.0, 300.1};
-  tiledb_datatype_t key3_type = TILEDB_FLOAT64;
-  uint64_t key3_size = 2 * sizeof(double);
-
-  char key4[] = "key_4";
-  const char* key5 = "invalid";
-  tiledb_datatype_t key4_type = TILEDB_CHAR;
-  uint64_t key4_size = strlen(key4) + 1;
-
-  // Get key-value item #1
-  tiledb_kv_item_t* kv_item1;
-  rc = tiledb_kv_get_item(ctx_, kv, &key1, key1_type, key1_size, &kv_item1);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item1, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY1_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-
-  // Attempt to get value for a2 (should give error)
-  rc =
-      tiledb_kv_item_get_value(ctx_, kv_item1, ATTR_2, &a2, &a2_type, &a2_size);
-  CHECK(rc == TILEDB_ERR);
-
-  // Get key-value item #2
-  tiledb_kv_item_t* kv_item2;
-  rc = tiledb_kv_get_item(ctx_, kv, &key2, key2_type, key2_size, &kv_item2);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item2, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY2_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-
-  // Get key-value item #3
-  tiledb_kv_item_t* kv_item3;
-  rc = tiledb_kv_get_item(ctx_, kv, key3, key3_type, key3_size, &kv_item3);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item3, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY3_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-
-  // Get key-value item #4
-  tiledb_kv_item_t* kv_item4;
-  rc = tiledb_kv_get_item(ctx_, kv, key4, key4_type, key4_size, &kv_item4);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item4, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY4_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-
-  // Invalid key
-  tiledb_kv_item_t* kv_item5;
-  rc = tiledb_kv_get_item(ctx_, kv, key5, key4_type, key4_size, &kv_item5);
-  REQUIRE(rc == TILEDB_OK);
-  CHECK(kv_item5 == nullptr);
-
-  // Attempt to write (should give error)
-  tiledb_kv_item_t* kv_item6;
-  rc = tiledb_kv_item_alloc(ctx_, &kv_item6);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_add_item(ctx_, kv, kv_item6);
-  CHECK(rc == TILEDB_ERR);
-
-  // Close key-value store
-  rc = tiledb_kv_close(ctx_, kv);
-  REQUIRE(rc == TILEDB_OK);
-
-  // Clean up
-  tiledb_kv_free(&kv);
-  tiledb_kv_item_free(&kv_item1);
-  tiledb_kv_item_free(&kv_item2);
-  tiledb_kv_item_free(&kv_item3);
-  tiledb_kv_item_free(&kv_item4);
-}
-
-void KVFx::check_interleaved_read_write(const std::string& path) {
-  // Open the key-value store
-  tiledb_kv_t* kv;
-  int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, nullptr, 0);
-  REQUIRE(rc == TILEDB_OK);
-
-  // Add an item
-  int new_key = 123;
-  tiledb_kv_item_t* kv_item1;
-  rc = tiledb_kv_item_alloc(ctx_, &kv_item1);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_key(
-      ctx_, kv_item1, &new_key, TILEDB_INT32, sizeof(int));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_1.c_str(), &KEY1_A1, TILEDB_INT32, sizeof(int));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_2, KEY1_A2, TILEDB_CHAR, strlen(KEY1_A2) + 1);
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_3, KEY1_A3, TILEDB_FLOAT32, 2 * sizeof(float));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_add_item(ctx_, kv, kv_item1);
-  CHECK(rc == TILEDB_OK);
-
-  // Read the new item
-  tiledb_kv_item_t* kv_item2;
-  const void *a1, *a2, *a3;
-  tiledb_datatype_t a1_type, a2_type, a3_type;
-  uint64_t a1_size, a2_size, a3_size;
-  rc = tiledb_kv_get_item(
-      ctx_, kv, &new_key, TILEDB_INT32, sizeof(int), &kv_item2);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item1, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY1_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-  rc =
-      tiledb_kv_item_get_value(ctx_, kv_item1, ATTR_2, &a2, &a2_type, &a2_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(!strcmp((const char*)a2, KEY1_A2));
-  CHECK(a2_type == TILEDB_CHAR);
-  CHECK(a2_size == strlen(KEY1_A2) + 1);
-  rc =
-      tiledb_kv_item_get_value(ctx_, kv_item1, ATTR_3, &a3, &a3_type, &a3_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(!memcmp(a3, KEY1_A3, 2 * sizeof(float)));
-  CHECK(a3_type == TILEDB_FLOAT32);
-  CHECK(a3_size == 2 * sizeof(float));
-
-  // Flush
-  rc = tiledb_kv_flush(ctx_, kv);
-  REQUIRE(rc == TILEDB_OK);
-
-  // Read the new item again
-  tiledb_kv_item_t* kv_item3;
-  rc = tiledb_kv_get_item(
-      ctx_, kv, &new_key, TILEDB_INT32, sizeof(int), &kv_item3);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_get_value(
-      ctx_, kv_item3, ATTR_1.c_str(), &a1, &a1_type, &a1_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(*(int*)a1 == KEY1_A1);
-  CHECK(a1_type == TILEDB_INT32);
-  CHECK(a1_size == sizeof(int));
-  rc =
-      tiledb_kv_item_get_value(ctx_, kv_item3, ATTR_2, &a2, &a2_type, &a2_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(!strcmp((const char*)a2, KEY1_A2));
-  CHECK(a2_type == TILEDB_CHAR);
-  CHECK(a2_size == strlen(KEY1_A2) + 1);
-  rc =
-      tiledb_kv_item_get_value(ctx_, kv_item3, ATTR_3, &a3, &a3_type, &a3_size);
-  CHECK(rc == TILEDB_OK);
-  CHECK(!memcmp(a3, KEY1_A3, 2 * sizeof(float)));
-  CHECK(a3_type == TILEDB_FLOAT32);
-  CHECK(a3_size == 2 * sizeof(float));
-
-  // Close key-value store
-  rc = tiledb_kv_close(ctx_, kv);
-  REQUIRE(rc == TILEDB_OK);
-
-  // Clean up
-  tiledb_kv_free(&kv);
-  tiledb_kv_item_free(&kv_item1);
-  tiledb_kv_item_free(&kv_item2);
-}
-
 void KVFx::check_kv_item(tiledb_kv_item_t* kv_item) {
   const void *key, *value;
   tiledb_datatype_t key_type, value_type;
@@ -927,14 +728,11 @@ void KVFx::check_kv_item(tiledb_kv_item_t* kv_item) {
 
 void KVFx::check_iter(const std::string& path) {
   // Create key-value iterator
-  const char* attributes[] = {ATTR_1.c_str(), ATTR_2, ATTR_3};
-
   tiledb_kv_t* kv;
   int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
   REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, attributes, 3);
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_READ);
   REQUIRE(rc == TILEDB_OK);
-
   tiledb_kv_iter_t* kv_iter;
   rc = tiledb_kv_iter_alloc(ctx_, kv, &kv_iter);
   REQUIRE(rc == TILEDB_OK);
@@ -971,46 +769,73 @@ void KVFx::check_iter(const std::string& path) {
   tiledb_kv_iter_free(&kv_iter);
 }
 
-void KVFx::check_kv_iter_with_dirty_kv(const std::string& path) {
-  // Create a kv and add an item
+TEST_CASE_METHOD(
+    KVFx, "C API: Test key-value mode", "[capi], [kv], [kv-mode]") {
+  std::string array_name;
+  create_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
+  array_name = FILE_URI_PREFIX + FILE_TEMP_DIR + KV_NAME;
+  create_kv(array_name);
+
+  // Open key-value store in WRITE mode
   tiledb_kv_t* kv;
-  int rc = tiledb_kv_alloc(ctx_, path.c_str(), &kv);
+  int rc = tiledb_kv_alloc(ctx_, array_name.c_str(), &kv);
   REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, nullptr, 0);
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_WRITE);
   REQUIRE(rc == TILEDB_OK);
 
-  // Add an item
-  int new_key = 123;
-  tiledb_kv_item_t* kv_item1;
-  rc = tiledb_kv_item_alloc(ctx_, &kv_item1);
-  REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_key(
-      ctx_, kv_item1, &new_key, TILEDB_INT32, sizeof(int));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_1.c_str(), &KEY1_A1, TILEDB_INT32, sizeof(int));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_2, KEY1_A2, TILEDB_CHAR, strlen(KEY1_A2) + 1);
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_item_set_value(
-      ctx_, kv_item1, ATTR_3, KEY1_A3, TILEDB_FLOAT32, 2 * sizeof(float));
-  CHECK(rc == TILEDB_OK);
-  rc = tiledb_kv_add_item(ctx_, kv, kv_item1);
-  CHECK(rc == TILEDB_OK);
+  char key[] = "key";
+  int has_key;
+  rc = tiledb_kv_has_key(ctx_, kv, key, TILEDB_CHAR, strlen(key), &has_key);
+  CHECK(rc == TILEDB_ERR);
 
-  // Create a kv iter with the dirty kv - should error
+  tiledb_kv_item_t* kv_item;
+  rc = tiledb_kv_get_item(ctx_, kv, &key, TILEDB_CHAR, strlen(key), &kv_item);
+  CHECK(rc == TILEDB_ERR);
+
+  rc = tiledb_kv_reopen(ctx_, kv);
+  CHECK(rc == TILEDB_ERR);
+
   tiledb_kv_iter_t* kv_iter;
   rc = tiledb_kv_iter_alloc(ctx_, kv, &kv_iter);
-  REQUIRE(rc == TILEDB_ERR);
+  CHECK(rc == TILEDB_ERR);
 
-  // Close and clean up
   rc = tiledb_kv_close(ctx_, kv);
+  CHECK(rc == TILEDB_OK);
+
+  // Open key-value store in READ mode
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_READ);
+  CHECK(rc == TILEDB_OK);
+
+  int dirty;
+  rc = tiledb_kv_is_dirty(ctx_, kv, &dirty);
+  CHECK(rc == TILEDB_ERR);
+
+  rc = tiledb_kv_flush(ctx_, kv);
+  CHECK(rc == TILEDB_ERR);
+
+  rc = tiledb_kv_item_alloc(ctx_, &kv_item);
   REQUIRE(rc == TILEDB_OK);
+  rc = tiledb_kv_item_set_key(ctx_, kv_item, key, TILEDB_CHAR, strlen(key));
+  REQUIRE(rc == TILEDB_OK);
+  int value_int = 0;
+  rc = tiledb_kv_item_set_value(
+      ctx_, kv_item, "attr", &value_int, TILEDB_INT32, sizeof(int));
+  REQUIRE(rc == TILEDB_OK);
+
+  rc = tiledb_kv_add_item(ctx_, kv, kv_item);
+  CHECK(rc == TILEDB_ERR);
+
+  // Clean up
+  rc = tiledb_kv_close(ctx_, kv);
+  CHECK(rc == TILEDB_OK);
   tiledb_kv_free(&kv);
+  tiledb_kv_item_free(&kv_item);
+
+  remove_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
 }
 
-TEST_CASE_METHOD(KVFx, "C API: Test key-value", "[capi], [kv]") {
+TEST_CASE_METHOD(
+    KVFx, "C API: Test key-value write, read and iter", "[capi], [kv]") {
   std::string array_name;
 
   if (supports_s3_) {
@@ -1020,10 +845,7 @@ TEST_CASE_METHOD(KVFx, "C API: Test key-value", "[capi], [kv]") {
     create_kv(array_name);
     check_write(array_name);
     check_single_read(array_name);
-    check_read_on_attribute_subset(array_name);
     check_iter(array_name);
-    check_interleaved_read_write(array_name);
-    check_kv_iter_with_dirty_kv(array_name);
     remove_temp_dir(S3_TEMP_DIR);
   } else if (supports_hdfs_) {
     create_temp_dir(HDFS_TEMP_DIR);
@@ -1031,10 +853,7 @@ TEST_CASE_METHOD(KVFx, "C API: Test key-value", "[capi], [kv]") {
     create_kv(array_name);
     check_write(array_name);
     check_single_read(array_name);
-    check_read_on_attribute_subset(array_name);
     check_iter(array_name);
-    check_interleaved_read_write(array_name);
-    check_kv_iter_with_dirty_kv(array_name);
     remove_temp_dir(HDFS_TEMP_DIR);
   } else {
     // File
@@ -1044,10 +863,7 @@ TEST_CASE_METHOD(KVFx, "C API: Test key-value", "[capi], [kv]") {
     check_kv_item();
     check_write(array_name);
     check_single_read(array_name);
-    check_read_on_attribute_subset(array_name);
     check_iter(array_name);
-    check_interleaved_read_write(array_name);
-    check_kv_iter_with_dirty_kv(array_name);
     remove_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
   }
 }
@@ -1064,9 +880,7 @@ TEST_CASE_METHOD(
   check_kv_item();
   check_write(array_name);
   check_single_read(array_name);
-  check_read_on_attribute_subset(array_name);
   check_iter(array_name);
-  check_interleaved_read_write(array_name);
   remove_temp_dir(FILE_URI_PREFIX + FILE_TEMP_DIR);
 }
 
@@ -1082,7 +896,7 @@ TEST_CASE_METHOD(
   tiledb_kv_t* kv;
   int rc = tiledb_kv_alloc(ctx_, array_name.c_str(), &kv);
   REQUIRE(rc == TILEDB_OK);
-  rc = tiledb_kv_open(ctx_, kv, nullptr, 0);
+  rc = tiledb_kv_open(ctx_, kv, TILEDB_READ);
   REQUIRE(rc == TILEDB_OK);
 
   // Get schema